refactor: 大范围重构 - 包名标准化 + 模块重命名 + 编译修复
- 移除 service-util 重命名为 common-util，所有引用同步更新
- 包名从 com.qrh.youshangdache.* 统一为 com.youshangdache.*
- Feign Client 从 model/feign/ 迁移至 service-client/system/，返回值去除 Result 包装
- service-client 依赖配置补充，解决 service-map/service-rules 等编译缺失
- 恢复误删的 DelayedMqConfig.java，修复 MQ 模块编译
- 修复 XxlJobInfo 与 BaseEntity 的字段冲突及类型不匹配
- 修复 service-system 控制器返回类型及枚举导入问题
- 修复 service-rules 残留 com.atguigu.daijia 包引用
- 父 POM 统一 <finalName> 为 ${project.artifactId}
- 补充 web-mgr 缺失的 spring-boot-starter-security 依赖

+++ service/service-dispatch/src/main/java/com/youshangdache/dispatch/xxl/client/XxlJobClient.java
@@ -77,7 +77,7 @@ public class XxlJobClient {
      */
     public Boolean startJob(Long jobId) {
         XxlJobInfo xxlJobInfo = new XxlJobInfo();
-        xxlJobInfo.setId(jobId.intValue());
+        xxlJobInfo.setId(jobId);
 
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
@@ -101,7 +101,7 @@ public class XxlJobClient {
      */
     public Boolean stopJob(Long jobId) {
         XxlJobInfo xxlJobInfo = new XxlJobInfo();
-        xxlJobInfo.setId(jobId.intValue());
+        xxlJobInfo.setId(jobId);
 
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
@@ -125,7 +125,7 @@ public class XxlJobClient {
      */
     public Boolean removeJob(Long jobId) {
         XxlJobInfo xxlJobInfo = new XxlJobInfo();
-        xxlJobInfo.setId(jobId.intValue());
+        xxlJobInfo.setId(jobId);
 
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
